fileio: read_full_stream() predict eof on regfile
When we know the size, we don't try read past EOF so feof() won't be
true.  If we know the size, we can predict EOF on the next iteration and
prevent it altogether, suppressing the unnecessary buffer enlargement
and also fixing the file size == READ_FULL_BYTES_MAX E2BIG edge case.

For non-regular-files we still need to discover the EOF, which as the
code is structured currently requires exponentialy more buffer space to
be allocated.