core: introduce UseRootFileSystemNamespace option
Allow user to specify that unit has to run in init's files system
namespace. Previously it was possible, iff none of the file system
namespace related option was set (but documentation gave no such
warranty). I wanted to have a formal and documented option which
forces such behavior. This option is incompatible with: PrivateTmp=,
PrivateDevices=, ProtectSystem=, ProtectHome=, ReadOnlyPaths=,
InaccessiblePaths=, ReadWritePaths=, MountFlags=, meaning it
cannot be set to true, if any of the listed options has non default
value.

Why this is useful?
We like a lot setup in which ssh + few other of our daemons run
in the root namespace. Thanks to that the mounts done by our
daemons are visible to users who ssh to a server to debug issues.
They are also visible to chef and our monitoring tools.
Of course this can also be achieved by having the mounts marked as
shared and not sharing the namespace. However, shared mounts
has proven problematic to us over time (especially how easy it is
to create cycles).
In the current state of systemd we can make ssh + our daemons run
in the root namespace, by making sure they do not use any of the
file system namespace related options. The dark side of this
approach is, how easy it is to break it. We recently specified
MountFlags=private and all worked nicely, until we realized
ssh cannot see the mounts done by our daemon (and our tools depend
on it). What is worse the documentation never promises that units
will run in root namespace, it just happens to be so.
As such we want to have an explicit option which forces unit
to not clone file system namespace from systemd, but stay in
the original one.