udev-builtin-blkid: Add --filter-superblock-type and --filter-partition-type
Add command-line options to udev's builtin blkid application that allow the
user to filter filesystem probing for a specific filesystem by superblock
type, and to filter filesystem probing by partition type.

ex: blkid --filter-supblock-type nilfs2
ex: blkid --filter-partition-type gpt

This change was motivated by blkid incorrectly identifying a device's
partition type as nilfs2 when it was looking at the whole device (/dev/mmcblk0).
This is because nilfs2 looks for superblocks at location 0x400, and at location
(((part_size-512)/8)*512). The data at the end of the disk in the backup
superblock was part of an ext4 partition (/dev/mmcblk0p25), but libblkid's
nilfs2 checking identified the data in the partition to have matched the
superblock magic (0x3434), but failed the CRC check. It resulted in the disk
being identified as having a nilfs2 partition with a bad checksum (SBBADCSUM),
and the udev data was tainted by values set by the nilfs2 probing routine,
affecting subsequent udev rules.