sysctl: write settings in order of declaration
Files are sorted lexicographically, and settings are read one by one, and the
ordered is remembered and used during setting.

This makes the order predictable and controlled, which is desirable for sysctl
settings which depend on one another.

Note: The only tricky case is when a later file overwrites the setting from an
earlier file. With this patch, the original "early" position is forgotten, and
the setting is applied "late". This is the most straightforward implementation,
although I can imagine that in some circumstances it might be better to keep
the order of first declaration. If you have a use case where this would matter,
speak up.

Fixes #4205.