execute: drop group priviliges before set up PAM
Previously we'd set up the GID early, before doing PAM, but the UID
after doing so. 096424d changed that. Now we set up PAM first, and
the UID+GID later. However, for the PAM stub we manually fixed up
the UID then, but still not the GID.
This commit moves setting up the PAM after set up the GID.

Fix #4238