Ensure mount unit goes into failed state when mount command times out
Fixes https://github.com/systemd/systemd/issues/4275.

The problem here is that while mount_dispatch_timer() sets m->result to
MOUNT_FAILURE_TIMEOUT, this is then overwritten by mount_sigchld_event() to
MOUNT_SUCCESS. This is because of

        if (is_clean_exit(code, status, NULL))
                f = MOUNT_SUCCESS;

and is_clean_exit() considers the SIGTERM status from the mount process a clean
exit. Then later on, we do

                if (f == MOUNT_SUCCESS)
                        mount_enter_mounted(m, f);

causing the mount unit to incorrectly reach "mounted" state.

The fix is to use the original m->result rather than "f". (Note that "f" is
assigned to m->result if the latter is MOUNT_SUCCESS, so we don't lose errors
indicated by "f" if the unit is still in a successful state.)